Configure production rewarded ad unit IDs
Replaces the test AdMob ad unit IDs with the official production IDs for rewarded ads on both iOS and Android.
This change configures the adUnitId in the adService to use the production identifiers when the application is running in a release environment (__DEV__ is false). The TestIds.REWARDED will continue to be used during development to ensure safe testing.
This is a necessary step to enable the serving of live, monetized ads to users in the production version of the app.